Added in basic functionality for automatic scene editing.
Currently, given a settlement id to link it to, it will take data in the csv to map it to the proper location.
All settlements that are not linked will be deleted.

It is a planned feature to be able to automatically create new settlements for the scene file, however that is still unifinished.